#137 provide AsyncFileChannelReader impl and benchmarks
