+++ reactor-extra/src/jmh/resources/file.txt
@@ -1,0 +1,10 @@
+1
+2
+3
+4
+5
+6
+7
+8
+9
+10 11 12
